+++ app/src/main/res/menu/menu_data_grid.xml
@@ -34,10 +34,10 @@
         app:showAsAction="always"
         android:title="@string/menu_action_heatmap_data_grid_title" />
     <item
-        android:id="@+id/menu_data_grid_action_trial_view"
+        android:id="@+id/menu_data_grid_action_map_view"
         android:icon="@drawable/ic_tb_toggle_all"
         app:showAsAction="ifRoom"
-        android:title="@string/menu_data_grid_action_trial_view"
+        android:title="@string/menu_data_grid_action_map_view"
         android:orderInCategory="1" />
     <item
         android:id="@+id/menu_data_grid_action_cell_details"
